pyproject.toml
[project]
name = "mcp-playwright-pytest-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openai-agents>=0.5.1",
    "pyotp>=2.9.0",
    "pytest>=8.4.2",
    "pytest-asyncio>=1.1.0",
    "pytest-html>=4.1.1",
]
